+++ src/camera/OrthographicCamera.h
@@ -6,8 +6,8 @@
 
 class OrthographicCamera : public Camera {
 public:
-    OrthographicCamera(Vector3 position, Vector3 direction, float viewportDistance, Vector3 up = Vector3(0, 1, 0));
-    Ray getRay(float xCenter, float yCenter) override;
+    OrthographicCamera(Vector3 position, Vector3 direction, double viewportDistance, Vector3 up = Vector3(0, 1, 0));
+    Ray getRay(double xCenter, double yCenter) override;
 };
