+++ src/camera/PerspectiveCamera.h
@@ -7,8 +7,8 @@
 
 class PerspectiveCamera : public Camera {
 public:
-    PerspectiveCamera(Vector3 position, Vector3 direction, float viewportDistance, Vector3 up = Vector3(0, 1, 0));
-    Ray getRay(float xCenter, float yCenter) override;
+    PerspectiveCamera(Vector3 position, Vector3 direction, double viewportDistance, Vector3 up = Vector3(0, 1, 0));
+    Ray getRay(double xCenter, double yCenter) override;
 };
