+++ src/geometry/Vector3.h
@@ -6,19 +6,19 @@
 
 class Vector3 {
 public:
-    float x = 0,
+    double x = 0,
           y = 0,
           z = 0;
 
     /// Creates vector between two points
     Vector3(const Vector3& p1, const Vector3& p2);
-    Vector3(float x, float y, float z);
-    Vector3(float xyz);
+    Vector3(double x, double y, double z);
+    Vector3(double xyz);
     Vector3() = default;
 
     Vector3 normalize() const;
-    float getMagnitude() const;
-    float dot(const Vector3 &other) const;
+    double getMagnitude() const;
+    double dot(const Vector3 &other) const;
     Vector3 cross(const Vector3 &other) const;
     Vector3 reflect(const Vector3& normal) const;
